+++ src/EightFace.cpp
@@ -669,10 +669,10 @@ struct EightFaceWidgetTemplate : ModuleWidget {
 		}
 
 		menu->addChild(new MenuSeparator());
-		menu->addChild(createSubmenuItem("Number of slots", "",
+		menu->addChild(createSubmenuItem("Number of slots", string::f("%i", module->presetCount),
 			[=](Menu* menu) {
 				menu->addChild(new NumberOfSlotsSlider(module));
-				menu->addChild(createBoolPtrMenuItem("Set by long press", "", &module->presetCountLongPress));
+				menu->addChild(createBoolPtrMenuItem("Set by long-press", "", &module->presetCountLongPress));
 			}
 		));